pyproject.toml
[tool.poetry]
name = "enprendete"
version = "0.1.0"
description = ""
authors = ["pdelsol <pato@admetricks.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.25.1"
python-dateutil = "^2.8.1"
ipdb = "^0.13.7"
pelican = {extras = ["markdown"], version = "^4.6.0"}
mutagen = "^1.45.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
